+++ src/bin/cmds/install.rs
@@ -39,12 +39,14 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
     let indices = c.indices.to_vec();
     let global_cache = c.directories.cache.clone();
     let threads = match_threads(c, args);
+    let shell = c.shell();
 
     let ctx = build::BuildCtx {
         indices,
         global_cache,
         logger,
         threads,
+        shell,
     };
 
     let targets = args