+++ src/bin/cmds/uninstall.rs
@@ -26,8 +26,9 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
         .unwrap_or_else(|| vec![]);
 
     let logger = logger(c);
+    let shell = c.shell();
 
-    let cache = Cache::from_disk(&logger, &c.directories.cache)?;
+    let cache = Cache::from_disk(&logger, &c.directories.cache, shell)?;
 
     let rc = cache.remove_bins(&spec, &targets)?;